Move to components insted of parallel routes.
