+++ app/(test)/taxonomy/page.tsx
@@ -1,11 +1,29 @@
+import { Suspense } from 'react';
+import FiltersServerComponent from './components/Filters/FiltersServerComponent';
+import ResultsServerComponent from './components/Results/ResultsServerComponent';
+
 export type PageProps = {
   children: React.ReactNode;
+  searchParams: URLSearchParams;
 };
 
-export const dynamic = 'force-dynamic';
+const Page = ({searchParams}:PageProps) => {
 
-const Page = ({children}:PageProps) => {
-  return children;
+  return (
+    <section>
+      <header>
+        <h2>Filters</h2>
+        <FiltersServerComponent searchParams={searchParams} />
+      </header>
+      <div>
+        <h2>Results</h2>
+        <Suspense fallback={<div>Loading...</div>}>
+          <ResultsServerComponent searchParams={searchParams} />
+        </Suspense>
+      </div>
+    </section>
+  );
+
 };
 
 export default Page;
